# BildixApi/bin/Debug/net9.0/openapi.yaml
...
openapi: 3.0.3
...
info:
  title: Bildix API
  version: 1.0.0
  description: |
    Recruitment platform API.
    Matches candidates with job offers using rule-based and semantic scoring.

# ── BLOC 2 : SERVEURS ─────────────────────────────────────────────────────────
# Où tourne l'API ? On peut en lister plusieurs (dev, staging, prod).
# Swagger UI propose une liste déroulante pour choisir contre quel serveur tester.
servers:
  - url: http://localhost:5000/api/v1
    description: Local development
  # - url: https://api.bildix.com/v1
  #   description: Production  (uncomment when deployed)

# ── BLOC 3 : SÉCURITÉ GLOBALE ─────────────────────────────────────────────────
# Déclare que TOUS les endpoints nécessitent un token JWT par défaut.
# Un endpoint peut surcharger cette règle avec "security: []" pour être public.
security:
  - bearerAuth: []

# ══════════════════════════════════════════════════════════════════════════════
# BLOC 4 : COMPOSANTS RÉUTILISABLES
# ══════════════════════════════════════════════════════════════════════════════
# Tout ce qui est défini ici peut être référencé ailleurs avec $ref.
# Exemple : $ref: '#/components/schemas/Candidate'
# Cela évite la répétition : on décrit une fois, on réutilise partout.
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ── Réponses d'erreur réutilisables ───────────────────────────────────────
  # Plutôt que répéter "401 Unauthorized" dans chaque endpoint, on le déclare
  # une fois ici et on y fait référence avec $ref.
  responses:
    Unauthorized:
      description: Missing or invalid JWT token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

  # ── Schémas de données ────────────────────────────────────────────────────
  # Un schéma décrit la FORME d'un objet JSON :
  # - quelles propriétés il a
  # - quel type chacune (string, integer, boolean, array, object…)
  # - lesquelles sont obligatoires (listed under "required")
  # - contraintes optionnelles (minLength, minimum, enum…)
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: NOT_FOUND
            message:
              type: string
              example: Candidate not found

    # ══════════════════════════════════════════════════════════════════════
    # CANDIDATE
    # ══════════════════════════════════════════════════════════════════════

    # Ce que l'API RETOURNE quand on lit un candidat.
    # Contient les champs calculés par le serveur (id, createdAt).
    Candidate:
      type: object
      required: [id, name, email, createdAt]
      properties:
        id:
          type: string
          format: uuid              # UUID v4 généré côté serveur
          example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
        name:
          type: string
          example: Alice Martin
        email:
          type: string
          format: email
          example: alice@example.com
        skills:
          type: array
          items:
            type: string
          example: [C#, SQL, Docker]
        location:
          type: string
          example: Paris, France
        experienceYears:
          type: integer
          minimum: 0
          example: 3
        createdAt:
          type: string
          format: date-time         # ISO 8601 : "2024-05-20T10:30:00Z"
          example: "2024-05-20T10:30:00Z"

    # Ce que le CLIENT ENVOIE pour créer un candidat.
    # Pas d'id ni de createdAt : c'est le serveur qui les génère.
    CreateCandidateRequest:
      type: object
      required: [name, email]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
          example: Alice Martin
        email:
          type: string
          format: email
          example: alice@example.com
        skills:
          type: array
          items:
            type: string
          example: [C#, SQL, Docker]
        location:
          type: string
          example: Paris, France
        experienceYears:
          type: integer
          minimum: 0
          example: 3

    # ══════════════════════════════════════════════════════════════════════
    # COMPANY
    # ══════════════════════════════════════════════════════════════════════
...
    Match:
      type: object
      required: [id, candidateId, jobOfferId, score, status, createdAt]
      properties:
        id:
          type: string
          format: uuid
        candidateId:
          type: string
          format: uuid
        jobOfferId:
          type: string
          format: uuid
        score:
          type: number
          format: float
          minimum: 0
          maximum: 100
          example: 87.5
        status:
          type: string
          # Only these 3 values are valid — enforced at API level
          enum: [pending, accepted, rejected]
          example: pending
        createdAt:
          type: string
          format: date-time
    UpdateMatchStatusRequest:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [accepted, rejected]

    # ══════════════════════════════════════════════════════════════════════
    # AUTH
    # ══════════════════════════════════════════════════════════════════════
...
paths:
...
  /candidates:
    # GET /api/v1/candidates → liste tous les candidats
    get:
      summary: List all candidates
      tags: [Candidates]           # Regroupe les endpoints dans Swagger UI
      responses:
        '200':
          description: Array of candidates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Candidate'
        '401':
          $ref: '#/components/responses/Unauthorized'

    # POST /api/v1/candidates → crée un nouveau candidat
    post:
      summary: Create a candidate profile
      tags: [Candidates]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCandidateRequest'
      responses:
        '201':                     # 201 Created (pas 200) pour une création
          description: Candidate created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Candidate'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /candidates/{id}:
    # {id} est un path parameter : /candidates/3fa85f64-...
    get:
      summary: Get a candidate by ID
      tags: [Candidates]
      parameters:
        - name: id
          in: path                 # "path" = dans l'URL, "query" = après le ?
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Candidate found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Candidate'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      summary: Delete a candidate
      tags: [Candidates]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':                     # 204 No Content : succès mais rien à retourner
          description: Candidate deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ── COMPANIES ──────────────────────────────────────────────────────────────
...
  /match/candidates/{jobId}:
    get:
      summary: Get top matching candidates for a job offer
      description: |
        Runs the scoring engine against all candidates and returns them ranked
        by match score (highest first). This is the core matching endpoint.
      tags: [Matching]
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: Ranked list of matches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Match'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /match/{candidateId}/{jobId}:
    post:
      summary: Compute and store a match between a candidate and a job offer
      tags: [Matching]
      parameters:
        - name: candidateId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: jobId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '201':
          description: Match computed and stored
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Match'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /match/{id}/status:
    patch:
      summary: Accept or reject a match (candidate or company action)
      tags: [Matching]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMatchStatusRequest'
      responses:
        '200':
          description: Match status updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Match'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ── AUTH ───────────────────────────────────────────────────────────────────
  # These endpoints are public (security: []) — they don't require a JWT token.
  # They ARE the endpoints used to obtain a token.
...